+++ rust/src/producers/from_r1cs.rs
@@ -3,6 +3,7 @@ use num_traits::One;
 use std::ops::Add;
 
 use crate::producers::builder::{BuildGate, GateBuilder, GateBuilderT};
+use crate::structs::relation::{ARITH, SIMPLE};
 use crate::{Header, Result, Sink, WireId};
 use BuildGate::*;
 
@@ -25,7 +26,12 @@ impl<S: Sink> FromR1CSConverter<S> {
     /// the ZKI CircuitHeader will be used to preallocate things
     pub fn new(sink: S, zki_header: &zkiCircuitHeader) -> Self {
         let mut conv = Self {
-            b: GateBuilder::new(sink, zki_header_to_header(zki_header).unwrap()),
+            b: GateBuilder::new(
+                sink,
+                zki_header_to_header(zki_header).unwrap(),
+                ARITH,
+                SIMPLE,
+            ),
             r1cs_to_ir_wire: Default::default(),
             minus_one: 0,
         };